requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
distlib==0.3.6
Django==4.1.4
django-money==3.0.0
django-moneyfield==0.2.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
filelock==3.9.0
frozenlist==1.3.3
game==0.0.5
github==1.2.7
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.0
Panda3D==1.10.13
panda3d-gltf==0.14
panda3d-simplepbr==0.10
Pillow==9.3.0
pipenv==2023.2.4
platformdirs==3.0.0
pluggy==1.0.0
py-moneyed==2.0
pygame==2.1.2
PyJWT==2.6.0
pyperclip==1.8.2
pytest==7.2.2
pytz==2022.7.1
sqlparse==0.4.3
style==1.1.0
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
update==0.0.1
ursina==5.2.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
yarl==1.9.2
